dark mode에서 theme error
