+++ src/App.css
@@ -36,3 +36,17 @@
 		transform: rotate(360deg);
 	}
 }
+
+html,
+body,
+#root {
+	margin: 0;
+	padding: 0;
+	height: 100%;
+	width: 100%;
+}
+
+body {
+	display: flex;
+	flex-direction: column;
+}
